chore(frontend): move SensitiveData helpers to React layer
Six framework-neutral SensitiveData helpers (types, utils, exemptionDataUtils,
components/utils, exemptionDataUtils.test, classification-example.json) had
zero non-React callers — only React pages and components imported them.

- Moved 6 files to react/lib/sensitive-data/
- Renamed components/utils.ts → components-utils.ts (flatten one level)
- Updated 13 React imports
- Removed empty src/components/SensitiveData/ directory
- Added `project.masking-exemption.level` to React locales (key was in
  Vue locale; exemptionDataUtils.ts now lives in React so the key needs
  to resolve via react-i18next)
- Extended no-react-to-vue-imports.test.ts ban list to include
  @/components/SensitiveData

One known cross-framework type leak remains in components-utils.ts:
`ResourceSelectOption` from @/components/v2/Select/RemoteResourceSelector/types.
That type ties to naive-ui's SelectOption and can't move until
@/components/ExprEditor/context's OptionConfig (which references it)
also migrates. Documented inline.
