+++ frontend/src/react/lib/sensitive-data/classification-example.json
