+++ frontend/src/react/lib/sensitive-data/exemptionDataUtils.test.ts
