+++ frontend/src/react/lib/sensitive-data/exemptionDataUtils.ts
