+++ frontend/src/react/lib/sensitive-data/types.ts
