+++ frontend/src/react/locales/es-ES.json
@@ -1712,6 +1712,7 @@
       "expires-in-days_other": "Expira en {{days}} días",
       "expires-today": "Expira hoy",
       "grant-exemption": "Conceder Exención",
+      "level": "nivel",
       "n-exemptions_one": "{{n}} exención",
       "n-exemptions_other": "{{n}} exenciones",
       "no-exemptions": "Aún no se han concedido exenciones de enmascaramiento",
