+++ frontend/src/react/locales/zh-CN.json
@@ -1712,6 +1712,7 @@
       "expires-in-days_other": "{{days}} 天后到期",
       "expires-today": "今天到期",
       "grant-exemption": "授予豁免",
+      "level": "等级",
       "n-exemptions_one": "{{n}} 项豁免",
       "n-exemptions_other": "{{n}} 项豁免",
       "no-exemptions": "尚未授予脱敏豁免",
